backend/requirements.txt
torch
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
fastapi
uvicorn[standard]
pymupdf
Pillow
numpy
scikit-learn
